+++ bit/cli/get_mapped_reads_pid.py
@@ -12,7 +12,7 @@ def build_parser():
         based on edit distance (using the NM field) and total alignment length. By default,
         it just prints out some summary stats. Specify an output file if you also want it to write out the
         percent identities for each mapped read. [bold]TO ALSO GET[/bold] coverage and detection information, use
-        `bit-get-cov-stats` instead. For version info, run `bit-version`.
+        `bit-cov-stats` instead. For version info, run `bit-version`.
         """
 
     parser = argparse.ArgumentParser(